Rename files
I find it is easier to keep editor tabs straight when there are fewer
files with the same name.
