+++ lib/Cake/Database/Dialect/SqliteDialectTrait.php
@@ -132,10 +132,10 @@ trait SqliteDialectTrait {
  * Used by Cake\Schema package to reflect schema and
  * generate schema.
  *
- * @return Cake\Database\Schema\Dialect\Sqlite
+ * @return Cake\Database\Schema\SqliteSchema
  */
 	public function schemaDialect() {
-		return new \Cake\Database\Schema\Dialect\Sqlite($this);
+		return new \Cake\Database\Schema\SqliteSchema($this);
 	}
 
 }
