+++ lib/Cake/Database/Schema/PostgresSchema.php
@@ -14,9 +14,9 @@
  * @since         CakePHP(tm) v 3.0.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace Cake\Database\Schema\Dialect;
+namespace Cake\Database\Schema;
 
-class Postgres {
+class PostgresSchema {
 
 /**
  * The driver instance being used.
