+++ lib/Cake/Test/TestCase/Database/Schema/MysqlSchemaTest.php
@@ -14,19 +14,19 @@
  * @since         CakePHP(tm) v 3.0.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace Cake\Test\TestCase\Database\Schema\Dialect;
+namespace Cake\Test\TestCase\Database\Schema;
 
 use Cake\Core\Configure;
 use Cake\Database\Connection;
 use Cake\Database\Schema\Collection as SchemaCollection;
-use Cake\Database\Schema\Dialect\Postgres;
-use Cake\Database\Schema\Driver\Postgres as PostgresDriver;
+use Cake\Database\Schema\MysqlSchema;
 use Cake\TestSuite\TestCase;
 
+
 /**
- * Postgres dialect test case.
+ * Test case for Mysql Schema Dialect.
  */
-class PostgresTest extends TestCase {
+class MysqlSchemaTest extends TestCase {
 
 /**
  * Helper method for skipping tests that need a real connection.
@@ -35,43 +35,7 @@ class PostgresTest extends TestCase {
  */
 	protected function _needsConnection() {
 		$config = Configure::read('Datasource.test');
-		$this->skipIf(strpos($config['datasource'], 'Postgres') === false, 'Not using Postgres for test config');
-	}
-
-/**
- * Helper method for testing methods.
- *
- * @return void
- */
-	protected function _createTables($connection) {
-		$this->_needsConnection();
-
-		$connection->execute('DROP TABLE IF EXISTS articles');
-		$connection->execute('DROP TABLE IF EXISTS authors');
-
-		$table = <<<SQL
-CREATE TABLE authors(
-id SERIAL,
-name VARCHAR(50),
-bio DATE,
-created TIMESTAMP
-)
-SQL;
-		$connection->execute($table);
-
-		$table = <<<SQL
-CREATE TABLE articles(
-id BIGINT PRIMARY KEY,
-title VARCHAR(20),
-body TEXT,
-author_id INTEGER NOT NULL,
-published BOOLEAN DEFAULT false,
-views SMALLINT DEFAULT 0,
-created TIMESTAMP
-)
-SQL;
-		$connection->execute($table);
-		$connection->execute('COMMENT ON COLUMN "articles"."title" IS \'a title\'');
+		$this->skipIf(strpos($config['datasource'], 'Mysql') === false, 'Not using Mysql for test config');
 	}
 
 /**
@@ -82,11 +46,7 @@ SQL;
 	public static function columnProvider() {
 		return [
 			[
-				'TIMESTAMP',
-				['datetime', null]
-			],
-			[
-				'TIMESTAMP WITHOUT TIME ZONE',
+				'DATETIME',
 				['datetime', null]
 			],
 			[
@@ -98,94 +58,128 @@ SQL;
 				['time', null]
 			],
 			[
-				'SMALLINT',
-				['integer', 5]
+				'TINYINT(1)',
+				['boolean', null]
 			],
 			[
-				'INTEGER',
-				['integer', 10]
+				'TINYINT(2)',
+				['integer', 2]
 			],
 			[
-				'SERIAL',
-				['integer', 10]
+				'INTEGER(11)',
+				['integer', 11]
 			],
 			[
 				'BIGINT',
-				['biginteger', 20]
+				['biginteger', null]
 			],
 			[
-				'NUMERIC',
-				['decimal', null]
+				'VARCHAR(255)',
+				['string', 255]
 			],
 			[
-				'DECIMAL(10,2)',
-				['decimal', null]
+				'CHAR(25)',
+				['string', 25]
 			],
 			[
-				'MONEY',
-				['decimal', null]
-			],
-			[
-				'VARCHAR',
+				'TINYTEXT',
 				['string', null]
 			],
 			[
-				'CHARACTER VARYING',
-				['string', null]
-			],
-			[
-				'CHAR',
-				['string', null]
-			],
-			[
-				'UUID',
-				['string', 36]
-			],
-			[
-				'CHARACTER',
-				['string', null]
-			],
-			[
-				'INET',
-				['string', 39]
-			],
-			[
-				'TEXT',
-				['text', null]
-			],
-			[
-				'BYTEA',
+				'BLOB',
 				['binary', null]
 			],
 			[
-				'REAL',
+				'MEDIUMBLOB',
+				['binary', null]
+			],
+			[
+				'FLOAT',
 				['float', null]
 			],
 			[
-				'DOUBLE PRECISION',
+				'DOUBLE',
 				['float', null]
 			],
 			[
-				'BIGSERIAL',
-				['biginteger', 20]
+				'DECIMAL(11,2)',
+				['decimal', null]
 			],
 		];
 	}
 
 /**
- * Test parsing Postgres column types.
+ * Test parsing MySQL column types.
  *
  * @dataProvider columnProvider
  * @return void
  */
 	public function testConvertColumnType($input, $expected) {
-		$driver = $this->getMock('Cake\Database\Driver\Postgres');
-		$dialect = new Postgres($driver);
+		$driver = $this->getMock('Cake\Database\Driver\Mysql');
+		$dialect = new MysqlSchema($driver);
 		$this->assertEquals($expected, $dialect->convertColumn($input));
 	}
 
 /**
- * Test listing tables with Postgres
+ * Provider for testing index conversion
+ *
+ * @return array
+ */
+	public static function convertIndexProvider() {
+		return [
+			['PRI', 'primary'],
+			['UNI', 'unique'],
+			['MUL', 'index'],
+		];
+	}
+/**
+ * Test parsing MySQL index types.
+ *
+ * @dataProvider convertIndexProvider
+ * @return void
+ */
+	public function testConvertIndex($input, $expected) {
+		$driver = $this->getMock('Cake\Database\Driver\Mysql');
+		$dialect = new MysqlSchema($driver);
+		$this->assertEquals($expected, $dialect->convertIndex($input));
+	}
+
+/**
+ * Helper method for testing methods.
+ *
+ * @return void
+ */
+	protected function _createTables($connection) {
+		$this->_needsConnection();
+		$connection->execute('DROP TABLE IF EXISTS articles');
+		$connection->execute('DROP TABLE IF EXISTS authors');
+
+		$table = <<<SQL
+CREATE TABLE authors(
+id INT(11) PRIMARY KEY AUTO_INCREMENT,
+name VARCHAR(50),
+bio TEXT,
+created DATETIME
+)
+SQL;
+		$connection->execute($table);
+
+		$table = <<<SQL
+CREATE TABLE articles(
+id BIGINT PRIMARY KEY AUTO_INCREMENT,
+title VARCHAR(20) COMMENT 'A title',
+body TEXT,
+author_id INT(11) NOT NULL,
+published BOOLEAN DEFAULT 0,
+allow_comments TINYINT(1) DEFAULT 0,
+created DATETIME
+) COLLATE=utf8_general_ci
+SQL;
+		$connection->execute($table);
+	}
+
+/**
+ * Integration test for SchemaCollection & MysqlDialect.
  *
  * @return void
  */
@@ -195,6 +189,7 @@ SQL;
 
 		$schema = new SchemaCollection($connection);
 		$result = $schema->listTables();
+
 		$this->assertInternalType('array', $result);
 		$this->assertCount(2, $result);
 		$this->assertEquals('articles', $result[0]);
@@ -202,7 +197,7 @@ SQL;
 	}
 
 /**
- * Test describing a table with Postgres
+ * Test describing a table with Mysql
  *
  * @return void
  */
@@ -212,6 +207,7 @@ SQL;
 
 		$schema = new SchemaCollection($connection);
 		$result = $schema->describe('articles');
+		$this->assertInstanceOf('Cake\Database\Schema\Table', $result);
 		$expected = [
 			'id' => [
 				'type' => 'biginteger',
@@ -228,9 +224,9 @@ SQL;
 				'null' => true,
 				'default' => null,
 				'length' => 20,
-				'comment' => 'a title',
+				'collate' => 'utf8_general_ci',
+				'comment' => 'A title',
 				'fixed' => null,
-				'collate' => null,
 				'charset' => null,
 			],
 			'body' => [
@@ -238,16 +234,16 @@ SQL;
 				'null' => true,
 				'default' => null,
 				'length' => null,
+				'collate' => 'utf8_general_ci',
 				'fixed' => null,
 				'comment' => null,
-				'collate' => null,
 				'charset' => null,
 			],
 			'author_id' => [
 				'type' => 'integer',
 				'null' => false,
 				'default' => null,
-				'length' => 10,
+				'length' => 11,
 				'fixed' => null,
 				'comment' => null,
 				'collate' => null,
@@ -263,11 +259,11 @@ SQL;
 				'collate' => null,
 				'charset' => null,
 			],
-			'views' => [
-				'type' => 'integer',
+			'allow_comments' => [
+				'type' => 'boolean',
 				'null' => true,
 				'default' => 0,
-				'length' => 5,
+				'length' => null,
 				'fixed' => null,
 				'comment' => null,
 				'collate' => null,
@@ -288,4 +284,5 @@ SQL;
 			$this->assertEquals($definition, $result->column($field));
 		}
 	}
+
 }
